requirements.txt
flask
requests
pyjwt
gunicorn==20.1.*